Replaced Poco's Timestamp with time_point from chrono library
* The time_points are updated using steady_clock to avoid problems of a non-monotonic system clock.
* Optimize time_point assignment when starting the periodic tasks

The variables logStatsTimePoint_ and readyLedBlinkTimePoint_ were
previously updated with two separate calls to chrono::steady_clock::now().
The compiler can not optimize calls to now() due to their nature.
Because these variables are updated in succession at start of the
ApiManager and do not depend on their execution order, they can be
updated with the value from a single same time_point.